+++ lib/src/routes/routes.dart
@@ -1,4 +1,5 @@
 import 'package:dreambook/src/ui/pages/shared/code_routes.dart';
+import 'package:dreambook/src/ui/pages/shared/settings_page.dart';
 import 'package:dreambook/src/ui/root.dart';
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
@@ -18,6 +19,7 @@ GoRouter appRoutes(AppRoutesRef ref) {
     TypedGoRoute<CupertinoRoute>(path: 'cupertino', name: 'cupertino'),
     TypedGoRoute<PaintingRoute>(path: 'painting', name: 'painting'),
     TypedGoRoute<BasicRoute>(path: 'basic', name: 'basic'),
+    TypedGoRoute<SettingsRoute>(path: 'settings', name: 'settings'),
   ],
 )
 class RootRoute extends GoRouteData {