+++ lib/src/routes/routes.g.dart
@@ -34,6 +34,11 @@ RouteBase get $rootRoute => GoRouteData.$route(
           name: 'basic',
           factory: $BasicRouteExtension._fromState,
         ),
+        GoRouteData.$route(
+          path: 'settings',
+          name: 'settings',
+          factory: $SettingsRouteExtension._fromState,
+        ),
       ],
     );
 
@@ -110,6 +115,23 @@ extension $BasicRouteExtension on BasicRoute {
 
   String get location => GoRouteData.$location(
         '/basic',
+      );
+
+  void go(BuildContext context) => context.go(location);
+
+  Future<T?> push<T>(BuildContext context) => context.push<T>(location);
+
+  void pushReplacement(BuildContext context) =>
+      context.pushReplacement(location);
+
+  void replace(BuildContext context) => context.replace(location);
+}
+
+extension $SettingsRouteExtension on SettingsRoute {
+  static SettingsRoute _fromState(GoRouterState state) => SettingsRoute();
+
+  String get location => GoRouteData.$location(
+        '/settings',
       );
 
   void go(BuildContext context) => context.go(location);