+++ lib/src/ui/pages/basic/text.g.dart
@@ -1,0 +1,24 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'text.dart';
+
+// **************************************************************************
+// RiverpodGenerator
+// **************************************************************************
+
+String _$configHash() => r'02f7ca125127d75745c46e5cf4623ca572b40992';
+
+/// See also [Config].
+@ProviderFor(Config)
+final configProvider = AutoDisposeNotifierProvider<Config, TextConfig>.internal(
+  Config.new,
+  name: r'configProvider',
+  debugGetCreateSourceHash:
+      const bool.fromEnvironment('dart.vm.product') ? null : _$configHash,
+  dependencies: null,
+  allTransitiveDependencies: null,
+);
+
+typedef _$Config = AutoDisposeNotifier<TextConfig>;
+// ignore_for_file: type=lint
+// ignore_for_file: subtype_of_sealed_class, invalid_use_of_internal_member, invalid_use_of_visible_for_testing_member