+++ lib/src/ui/pages/basic/widgets.dart
@@ -1,3 +1,7 @@
 import 'package:dreambook/src/ui/pages/shared/shared_code_view.dart';
 
-final basicCodeItems = <CodeItem>[];
+import 'text.dart';
+
+final basicCodeItems = <CodeItem>[
+  textItem,
+];